Mise à jour des view pour structuration par dossier : - Menu - questionnaire - login - home Création de la fonction toString() pour mcq et question pour affichage dans la liste des questions
