+++ my_qcm/app/src/main/java/com/tactfactory/my_qcm/entity/Question.java
@@ -88,4 +88,9 @@ public class Question {
     public void setUpdated_at(Date updated_at) {
         this.updated_at = updated_at;
     }
+
+        @Override
+    public String toString() {
+        return ques +" "+ mcq.toString();
+    }
 }
