+++ my_qcm/app/src/main/java/com/tactfactory/my_qcm/view/login/LoginActivity.java
@@ -1,9 +1,7 @@
-package com.tactfactory.my_qcm.view;
+package com.tactfactory.my_qcm.view.login;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
@@ -12,6 +10,7 @@ import android.view.MenuItem;
 import android.widget.Button;
 
 import com.tactfactory.my_qcm.R;
+import com.tactfactory.my_qcm.view.home.HomeActivity;
 
 public class LoginActivity extends AppCompatActivity {
