+++ my_qcm/app/src/main/java/com/tactfactory/my_qcm/view/menu/HelpFragment.java
@@ -1,4 +1,4 @@
-package com.tactfactory.my_qcm.view;
+package com.tactfactory.my_qcm.view.menu;
 
 
 import android.os.Bundle;
@@ -12,10 +12,10 @@ import com.tactfactory.my_qcm.R;
 /**
  * A simple {@link Fragment} subclass.
  */
-public class ProfileFragment extends Fragment {
+public class HelpFragment extends Fragment {
 
 
-    public ProfileFragment() {
+    public HelpFragment() {
         // Required empty public constructor
     }
 
@@ -24,7 +24,7 @@ public class ProfileFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_profile, container, false);
+        return inflater.inflate(R.layout.fragment_help, container, false);
     }
 
 }
