+++ my_qcm/app/src/main/res/layout/app_bar_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.tactfactory.my_qcm.view.HomeActivity">
+    tools:context=".view.home.HomeActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
