+++ my_qcm/app/src/main/res/layout/content_login.xml
@@ -2,7 +2,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/fragment"
-    android:name="com.tactfactory.my_qcm.view.LoginActivityFragment"
+    android:name="com.tactfactory.my_qcm.view.login.LoginActivityFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
