+++ my_qcm/app/src/main/res/layout/fragment_home.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.HomeFragment">
+    tools:context=".view.home.HomeFragment">
 
 
     <ListView
