+++ my_qcm/app/src/main/res/layout/fragment_legal_notices.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.LegalNoticesFragment">
+    tools:context=".view.menu.LegalNoticesFragment">
 
 
     <Button
