+++ my_qcm/app/src/main/res/layout/fragment_login.xml
@@ -6,7 +6,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".view.LoginActivityFragment"
+    tools:context=".view.login.LoginActivityFragment"
     tools:showIn="@layout/activity_login">
 
     <EditText
