+++ my_qcm/app/src/main/res/layout/fragment_profile.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.tactfactory.my_qcm.view.ProfileFragment">
+    tools:context=".view.menu.ProfileFragment">
