+++ my_qcm/app/src/main/res/layout/fragment_rules.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.tactfactory.my_qcm.view.RulesFragment">
+    tools:context=".view.menu.RulesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
